+++ examples/foxes/tab_file/run.py
@@ -1,0 +1,147 @@
+import time
+import argparse
+import matplotlib.pyplot as plt
+
+import foxes
+import foxes.variables as FV
+import foxes.constants as FC
+from foxes.utils.runners import DaskRunner
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "-l",
+        "--layout",
+        help="The wind farm layout file (path or static)",
+        default="test_farm_67.csv",
+    )
+    parser.add_argument(
+        "-s",
+        "--states",
+        help="The states input file (path or static)",
+        default="winds100.tab",
+    )
+    parser.add_argument(
+        "-t",
+        "--turbine_file",
+        help="The P-ct-curve csv file (path or static)",
+        default="NREL-5MW-D126-H90.csv",
+    )
+    parser.add_argument("-r", "--rotor", help="The rotor model", default="centre")
+    parser.add_argument(
+        "-p", "--pwakes", help="The partial wakes model", default="rotor_points"
+    )
+    parser.add_argument(
+        "-w",
+        "--wakes",
+        help="The wake models",
+        default=["CrespoHernandez_quadratic", "Bastankhah2014_linear"],
+        nargs="+",
+    )
+    parser.add_argument(
+        "-m", "--tmodels", help="The turbine models", default=[], nargs="+"
+    )
+    parser.add_argument(
+        "-sl",
+        "--show_layout",
+        help="Flag for showing layout figure",
+        action="store_true",
+    )
+    parser.add_argument(
+        "-cm", "--calc_mean", help="Calculate states mean", action="store_true"
+    )
+    parser.add_argument(
+        "-c", "--chunksize", help="The maximal chunk size", type=int, default=1000
+    )
+    parser.add_argument("-sc", "--scheduler", help="The scheduler choice", default=None)
+    parser.add_argument(
+        "-n",
+        "--n_workers",
+        help="The number of workers for distributed run",
+        type=int,
+        default=None,
+    )
+    parser.add_argument(
+        "-tw",
+        "--threads_per_worker",
+        help="The number of threads per worker for distributed run",
+        type=int,
+        default=None,
+    )
+    parser.add_argument(
+        "--nodask", help="Use numpy arrays instead of dask arrays", action="store_true"
+    )
+    args = parser.parse_args()
+
+    cks = None if args.nodask else {FC.STATE: args.chunksize}
+    if args.calc_mean:
+        cks[FC.POINT] = 4000
+
+    mbook = foxes.models.ModelBook()
+    ttype = foxes.models.turbine_types.PCtFile(args.turbine_file)
+    mbook.turbine_types[ttype.name] = ttype
+
+    states = foxes.input.states.TabStates(
+        data_source=args.states,
+        output_vars=[FV.WS, FV.WD, FV.TI, FV.RHO],
+        fixed_vars={FV.RHO: 1.225, FV.TI: 0.05},
+    )
+
+    o = foxes.output.StatesRosePlotOutput(states, point=[0.0, 0.0, 100.0])
+    fig = o.get_figure(12, FV.AMB_WS, [0, 3.5, 6, 10, 15, 20])
+    plt.show()
+
+    farm = foxes.WindFarm()
+    foxes.input.farm_layout.add_from_file(
+        farm,
+        args.layout,
+        col_x="x",
+        col_y="y",
+        col_H="H",
+        turbine_models=[ttype.name, "kTI_02"] + args.tmodels,
+    )
+
+    if args.show_layout:
+        ax = foxes.output.FarmLayoutOutput(farm).get_figure()
+        plt.show()
+        plt.close(ax.get_figure())
+
+    algo = foxes.algorithms.Downwind(
+        mbook,
+        farm,
+        states=states,
+        rotor_model=args.rotor,
+        wake_models=args.wakes,
+        wake_frame="rotor_wd",
+        partial_wakes_model=args.pwakes,
+        chunks=cks,
+    )
+
+    with DaskRunner(
+        scheduler=args.scheduler,
+        n_workers=args.n_workers,
+        threads_per_worker=args.threads_per_worker,
+    ) as runner:
+        time0 = time.time()
+        farm_results = runner.run(algo.calc_farm)
+        time1 = time.time()
+
+        print("\nCalc time =", time1 - time0, "\n")
+
+        print(farm_results)
+
+        fr = farm_results.to_dataframe()
+        print(fr[[FV.WD, FV.H, FV.AMB_REWS, FV.REWS, FV.AMB_P, FV.P, FV.WEIGHT]])
+
+        o = foxes.output.FarmResultsEval(farm_results)
+        P0 = o.calc_mean_farm_power(ambient=True)
+        P = o.calc_mean_farm_power()
+        print(f"\nFarm power        : {P/1000:.1f} MW")
+        print(f"Farm ambient power: {P0/1000:.1f} MW")
+        print(f"Farm efficiency   : {o.calc_farm_efficiency()*100:.2f} %")
+        print(f"Annual farm yield : {o.calc_farm_yield(algo=algo):.2f} GWh")
+
+        if args.calc_mean:
+            o = foxes.output.FlowPlots2D(algo, farm_results, runner=runner)
+            fig = o.get_mean_fig_xy(FV.WS, resolution=30)
+            plt.show()