+++ foxes/input/states/__init__.py
@@ -4,7 +4,7 @@ Atmospheric input states.
 
 from .single import SingleStateStates
 from .scan_ws import ScanWS
-from .states_table import StatesTable, Timeseries
+from .states_table import StatesTable, Timeseries, TabStates
 from .field_data_nc import FieldDataNC
 from .multi_height import MultiHeightStates, MultiHeightTimeseries